# backend-go/api/openapi.yaml
openapi: 3.0.3
info:
  title: xiaoan-platform API
  version: 2.0.0
  description: |
    xiaoan-platform backend contract (Go control plane).

    This file is the single source of truth for the HTTP API.
    Go server types and TypeScript clients are generated from it.

    ## Error envelope conventions (legacy-compatible)

    The React frontend was validated against the Django reference
    implementation, which produced four distinct error shapes:

    - `{"detail": "message"}` — 403/404 style errors (DRF PermissionDenied / NotFound).
    - `{"error": "message"}` — simple JSON errors (identity endpoints, 409/502 conflicts).
    - `{"field": ["message", ...]}` — field validation errors (DRF serializer style).
    - `["message"]` — bare array, 400 for hand-raised DRF ValidationErrors (e.g. "application_id is required").

    All four shapes are reproduced exactly. Clients should read the
    `detail` / `error` / first field message in that order.

    ## Authentication

    Normal users authenticate with the HttpOnly opaque session cookie
    `studio_session` (set by OAuth exchange / admin login). All
    state-changing requests additionally require the CSRF header
    `X-CSRF-Token` matching the `studio_csrf` cookie (double-submit).
    `GET /api/v2/artifacts/{id}/open` is cookie-authenticated because it is
    opened via browser navigation (window.open) and cannot send headers.

    ## SSE

    `GET /api/v2/runs/{id}/stream` responds with `text/event-stream`.
...
servers:
...
paths:
...
  /api/identity/oauth/start:
    get:
      operationId: oauthStart
      tags: [identity]
      summary: Redirect (302) the browser to the Feishu OAuth authorize page
      parameters:
        - name: return_to
          in: query
          schema: { type: string, default: "/" }
          description: SPA path to return to after login; must start with "/" and not "//".
      responses:
        "302":
          description: Redirect to Feishu authorize URL (state signed, 10 min TTL).
          headers:
            Location:
              schema: { type: string }
        "500":
          description: OAuth not configured
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleError" }
  /api/identity/oauth/exchange:
    get:
      operationId: oauthExchange
      tags: [identity]
      summary: Exchange the OAuth ?code&state for a studio session
      description: |
        Verifies the signed state, exchanges the code at the Feishu token
        endpoint, upserts the local user (auth_source=feishu), stores the
        refresh token AES-256-GCM encrypted, sets the `studio_session`
        HttpOnly cookie and returns the session user plus JWT tokens for the
        legacy transition window.
      parameters:
        - name: code
          in: query
          required: true
          schema: { type: string }
        - name: state
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Authenticated
          headers:
            Set-Cookie:
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                required: [user, tokens, return_to]
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  tokens: { $ref: "#/components/schemas/Tokens" }
                  return_to: { type: string }
        "400":
          description: missing code / invalid or expired oauth state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleError" }
        "502":
          description: oauth exchange failed / no access token / user info failed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleError" }
  /api/identity/session:
    get:
      operationId: getSession
      tags: [identity]
      summary: Current session user (cookie-authenticated)
      responses:
        "200":
          description: Session user payload
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionUser" }
        "401":
          description: not authenticated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleError" }
  /api/identity/admin/login:
    post:
      operationId: adminLogin
      tags: [identity]
      summary: Local admin login (Argon2id password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        "200":
          description: Admin session established
          content:
            application/json:
              schema:
                type: object
                required: [id, username, is_staff, auth_source]
                properties:
                  id: { type: integer }
                  username: { type: string }
                  is_staff: { type: boolean }
                  auth_source: { type: string }
                  display_name: { type: string }
                  display_id: { type: string }
        "401":
          description: invalid credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SimpleError" }

  # ─────────────────────────────────────────────────────────────── auth ──
  /api/auth/login/:
    post:
      operationId: authLogin
      tags: [auth]
      summary: Local username/password login (transition window)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        "200":
          description: Logged in; also sets studio_session cookie
          content:
            application/json:
              schema:
                type: object
                required: [user, tokens]
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  tokens: { $ref: "#/components/schemas/Tokens" }
        "401":
          description: bad credentials
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DetailError" }
...
  /api/auth/token/refresh/:
    post:
      operationId: tokenRefresh
      tags: [auth]
      summary: Rotate a JWT refresh token (transition window)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh]
              properties:
                refresh: { type: string }
      responses:
        "200":
          description: New access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  access: { type: string }
                  refresh: { type: string }
  /api/auth/me/:
    get:
      operationId: getMe
      tags: [auth]
      summary: Current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
    patch:
      operationId: patchMe
      tags: [auth]
      summary: Update current user profile
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username: { type: string }
                email: { type: string }
                avatar: { type: string }
                bio: { type: string }
                display_id: { type: string }
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }

  # ────────────────────────────────────────────── workspace bootstrap (P1-1) ──
  #
  # The one request every shell surface may use to learn the small,
  # constant-size facts it needs at start-up: which agent is the workspace
  # default, the home shortcut groups, and the category rails. It exists to
  # REPLACE the whole-catalog download (执行报告 §9–§11): the response is
  # ~30 rows no matter whether the database holds 7 or 50 000 applications.
  #
  # The application rows here are ApplicationSummary (display fields only) —
  # EXCEPT default_application, which additionally carries `skills` and
  # `capabilities` because the composer bound to it renders its 技能 chips
  # and gates the attachment entry on the runtime capabilities.
...
components:
...
  schemas:
...
    DetailError:
      type: object
      required: [detail]
      properties:
        detail: { type: string }
    SimpleError:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    Tokens:
      type: object
      properties:
        access: { type: string }
        refresh: { type: string }
    User:
      type: object
      description: session_user_payload — the single canonical user shape
      required: [id, username]
      properties:
        id: { type: integer }
        username: { type: string }
        email: { type: string }
        role: { type: string }
        avatar: { type: string }
        avatar_url: { type: string }
        auth_source: { type: string }
        is_staff: { type: boolean }
        display_name: { type: string }
        display_id: { type: string }
    SessionUser:
      type: object
      required: [username]
      properties:
        id: { type: integer }
        username: { type: string }
        display_name: { type: string }
        display_id: { type: string }
        avatar_url: { type: string }
        auth_source: { type: string }
        is_staff: { type: boolean }
...